### 2025-04-23 #7523 @samhforbes
Hi @genisplaja, @yiitozer, @olivecha would you be interested in reviewing this software for JOSS?
