### 2025-04-24 #8102 @editorialbot
[REVIEW]: Cosmodoit: A Python Package for Adaptive, Efficient Pipelining of Feature Extraction from Performed Music
**Submitting author:** @CGuichaoua (<a href="http://orcid.org/0000-0003-2432-2162">Corentin Guichaoua</a>)
**Repository:** https://github.com/erc-cosmos/cosmodoit/
**Branch with paper.md** (empty if default branch): Paper
**Version:** 0.5.0
