### 2026-02-21 #8102 @elaine-chew
Hi @samhforbes,

Apologies for this late response.  I am in the midst of the end-of-project
audit for the ERC grant that funded this research and on leave for a major
family event, and Daniel Bedoya (our other co-author) is also at the end of
another ERC project.  All are scheduled to finish at the end of February
after which we can pick this up.  Any chance we can re-open this to
continue the process without having to start from scratch all over again?

Best wishes,
Elaine
